+++ src/main/java/racingcar/controller/Controller.java
@@ -1,7 +1,7 @@
 package racingcar.controller;
 
 import java.util.List;
-import racingcar.domain.RacingGame;
+import racingcar.domain.game.RacingGame;
 import racingcar.domain.random.RandomNumberGenerator;
 import racingcar.parser.CarNameParser;
 import racingcar.parser.TryCountParser;
