+++ src/main/java/racingcar/domain/game/RacingGame.java
@@ -1,6 +1,8 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
 import java.util.List;
+import racingcar.domain.Cars;
+import racingcar.domain.TryCount;
 import racingcar.domain.random.RandomNumberGenerator;
 import racingcar.exception.GetWinnersBeforeFinishException;
 import racingcar.exception.RacingGameIsFinishedException;
